Change cache key to take 'lang' into account
The same source can pull in page data from multiple languages, but
the page name can potentially be the same. When we cache the results,
we should cache alongside the language requested.

Each source separately caches itself, so the only dynamically set
property is lang, and we should take that into account when creating
cache keys.